CMSIS-NN: Fix warnings
1. Doxygen warnings are fixed

2. Sign conversion warnings(-Wsign-conversion) and -Wextra warnings
   are fixed for TFLu compliant MVE APIs

+++ CMSIS/NN/Source/ConvolutionFunctions/arm_depthwise_conv_u8_basic_ver1.c
@@ -21,8 +21,8 @@
  * Title:        arm_depthwise_conv_u8_basic_ver1.c
  * Description:  u8 depthwise convolution function
  *
- * $Date:        May 8, 2020
- * $Revision:    V.1.0.0
+ * $Date:        May 29, 2020
+ * $Revision:    V.1.1.0
  *
  * Target :  Cortex-M CPUs
  *
@@ -128,24 +128,24 @@ static void depthwise_conv_u8_mult_4(const uint8_t *input,
 }
 
 static void depthwise_conv_u8_generic(const uint8_t *input,
-                                      const uint16_t input_x,
-                                      const uint16_t input_y,
-                                      const uint16_t input_ch,
+                                      const int32_t input_x,
+                                      const int32_t input_y,
+                                      const int32_t input_ch,
                                       const uint8_t *kernel,
-                                      const uint16_t output_ch,
-                                      const uint16_t ch_mult,
-                                      const uint16_t kernel_x,
-                                      const uint16_t kernel_y,
-                                      const uint16_t pad_x,
-                                      const uint16_t pad_y,
-                                      const uint16_t stride_x,
-                                      const uint16_t stride_y,
+                                      const int32_t output_ch,
+                                      const int32_t ch_mult,
+                                      const int32_t kernel_x,
+                                      const int32_t kernel_y,
+                                      const int32_t pad_x,
+                                      const int32_t pad_y,
+                                      const int32_t stride_x,
+                                      const int32_t stride_y,
                                       const int32_t *bias,
                                       uint8_t *output,
                                       const int32_t output_shift,
                                       const int32_t output_mult,
-                                      const uint16_t output_x,
-                                      const uint16_t output_y,
+                                      const int32_t output_x,
+                                      const int32_t output_y,
                                       const int32_t output_offset,
                                       const int32_t input_offset,
                                       const int32_t filter_offset,
@@ -231,8 +231,8 @@ static void depthwise_conv_u8_generic(const uint8_t *input,
  * @param[in]     output_y  Height of output tensor
  * @param[in]     output_activation_min   Minimum value to clamp the output to. Range : {0, 255}
  * @param[in]     output_activation_max   Minimum value to clamp the output to. Range : {0, 255}
- * @param[in]     out_shift  Amount of right-shift for output
- * @param[in]     out_mult   Output multiplier for requantization
+ * @param[in]     output_shift  Amount of right-shift for output
+ * @param[in]     output_mult   Output multiplier for requantization
  * @return        The function returns one of the following
  *                <code>ARM_MATH_SIZE_MISMATCH</code> - Not supported dimension of tensors
  *                <code>ARM_MATH_SUCCESS</code> - Successful operation
