+++ CMSIS/NN/Source/ConvolutionFunctions/arm_nn_mat_mult_kernel_s8_s16.c
@@ -21,8 +21,8 @@
  * Title:        arm_nn_mat_mult_kernel_s8_s16.c
  * Description:  Matrix-multiplication function for convolution
  *
- * $Date:        February 27, 2020
- * $Revision:    V.1.0.1
+ * $Date:        May 29, 2020
+ * $Revision:    V.1.0.2
  *
  * Target Processor:  Cortex-M cores
  * -------------------------------------------------------------------- */
@@ -173,7 +173,7 @@ q7_t *arm_nn_mat_mult_kernel_s8_s16(const q7_t *input_a,
     if (row_count)
     {
         ip_a0_s8 = input_a + num_col_a * (output_ch & ~3);
-        const mve_pred16_t p = vctp32q(row_count);
+        const mve_pred16_t p = vctp32q((uint32_t)row_count);
         int32x4_t out_vec_0 = vdupq_n_s32(0);
         int32x4_t out_vec_1 = vdupq_n_s32(0);
         int32x4_t mult_tail;
